fix(sslictcl): honour facts in policy, shared list parsing, reserved grade id, key-bits domain
Four defects from the review of #1771.

`evaluate_policy` ignored the declared catalogue. It passed `None` to
`cipher_has_forward_secrecy`, so a cipher declared `forward-secrecy false`
still passed `require-forward-secrecy` on the ECDHE name heuristic, and a
nonstandard suite declared `true` failed it — the policy and the estimator
could reach opposite conclusions about the same cipher from the same
document. `evaluate_policy` now takes the same `Option<&TlsFacts>` the
estimator does and threads it through, with a test covering both directions.

`literal_list` re-parsed a braced LIST as a *command*. That rejected
`forbid-ciphers {[A-Z]*RC4}` as command substitution, split on `;`, and
refused a list spanning lines. Lists are owned by `tcl_syntax::list` per
AGENTS.md's owner map, so a braced value now goes through `split_list`
(`Tcl_SplitList`): a braced word is verbatim by Tcl's own rules, so nothing
inside it needs a substitution check, and a malformed list is SSLIC1009. An
unbraced word carrying substitution is still SSLIC1002.

`check grade { … }` collided with the grade rule's finding identity — both
would have produced `(check_id = "grade", endpoint)`. `grade` is now
reserved as a check identifier (`dsl::RESERVED_CHECK_ID`), reported as
SSLIC1009, and said so in the vocabulary reference and in `vocabulary.rs`.

`min-key-bits` narrowed its parsed `u64` with `u32::try_from(...).ok()`, so
a bound above `u32::MAX` silently became `None` and the check passed
everything. `PolicyCheck::min_key_bits` is a `u64` and the comparison widens
the certificate's key size instead.

+++ rust/tcl-sslictcl/src/model.rs
@@ -404,9 +404,11 @@ pub struct PolicyCheck {
     /// Whether every enabled cipher must provide forward secrecy.
     #[serde(default)]
     pub require_forward_secrecy: Option<bool>,
-    /// Minimum leaf public-key size in bits.
+    /// Minimum leaf public-key size in bits. Held at the full `INT` width so a
+    /// value above `u32::MAX` still fails every real certificate instead of
+    /// being silently dropped.
     #[serde(default)]
-    pub min_key_bits: Option<u32>,
+    pub min_key_bits: Option<u64>,
     /// Whether HSTS must be enabled.
     #[serde(default)]
     pub require_hsts: Option<bool>,
